split mint_basic
