+++ crates/competitive/src/num/mint.rs
@@ -1,11 +1,11 @@
 //! modint
 
-#[codesnip::skip]
 use crate::{
     num::{One, Zero},
     tools::IterScan,
 };
 
+#[codesnip::entry("MIntBase", include("scanner", "zero_one"))]
 #[repr(transparent)]
 pub struct MInt<M>
 where
@@ -15,6 +15,7 @@ where
     _marker: std::marker::PhantomData<fn() -> M>,
 }
 
+#[codesnip::entry("MIntBase")]
 pub trait MIntBase {
     type Inner: Sized + Copy + Eq + std::fmt::Debug + std::hash::Hash;
     fn get_mod() -> Self::Inner;
@@ -39,28 +40,18 @@ pub trait MIntBase {
     }
 }
 
+#[codesnip::entry("MIntBase")]
 pub trait MIntConvert<T = <Self as MIntBase>::Inner>: MIntBase {
     fn from(x: T) -> <Self as MIntBase>::Inner;
     fn into(x: <Self as MIntBase>::Inner) -> T;
     fn mod_into() -> T;
 }
 
-pub mod mint_base;
+#[cfg_attr(nightly, codesnip::entry("MIntBase"))]
+mod mint_base;
 
-#[cfg_attr(nightly, codesnip::skip)]
+#[cfg_attr(nightly, codesnip::entry("MInt", include("MIntBase")))]
+pub mod mint_basic;
+
+#[cfg_attr(nightly, codesnip::entry("montgomery", include("MIntBase")))]
 pub mod montgomery;
-
-#[test]
-fn test_mint() {
-    use super::mint_base::MInt998244353;
-    use crate::tools::Xorshift;
-    let mut rand = Xorshift::default();
-    const Q: usize = 10_000;
-    for _ in 0..Q {
-        let a = MInt998244353::new(rand.rand(MInt998244353::get_mod() as u64 - 1) as u32 + 1);
-        let x = a.inv();
-        assert!(x.x < MInt998244353::get_mod());
-        assert_eq!(a * x, MInt998244353::one());
-        assert_eq!(x, a.pow(MInt998244353::get_mod() as usize - 2));
-    }
-}
