+++ crates/library_checker/src/datastructure/point_set_range_composite.rs
@@ -3,7 +3,7 @@ use competitive::prelude::*;
 pub use competitive::{
     algebra::LinearOperation,
     data_structure::SegmentTree,
-    num::{mint_base::MInt998244353, MInt},
+    num::{mint_basic::MInt998244353, MInt},
 };
 
 #[verify::verify("https://judge.yosupo.jp/problem/point_set_range_composite")]
