+++ crates/library_checker/src/datastructure/queue_operate_all_composite.rs
@@ -3,7 +3,7 @@ use competitive::prelude::*;
 pub use competitive::{
     algebra::LinearOperation,
     data_structure::QueueAggregation,
-    num::{mint_base::MInt998244353, MInt},
+    num::{mint_basic::MInt998244353, MInt},
 };
 
 #[verify::verify("https://judge.yosupo.jp/problem/queue_operate_all_composite")]
