+++ crates/library_checker/src/datastructure/vertex_set_path_composite.rs
@@ -4,7 +4,7 @@ pub use competitive::{
     algebra::{LinearOperation, ReverseOperation},
     data_structure::SegmentTree,
     graph::TreeGraphScanner,
-    num::{mint_base::MInt998244353, MInt},
+    num::{mint_basic::MInt998244353, MInt},
     tree::HeavyLightDecomposition,
 };
